+++ src/web/assets/cp/src/form-builder/formie-form-builder.js
@@ -84,7 +84,9 @@ import { FieldSelectField } from '@form-builder/fields/FieldSelectField';
 import { FieldConditionsField } from '@form-builder/fields/FieldConditionsField';
 import { NextButtonConditionsField } from '@form-builder/fields/NextButtonConditionsField';
 import { StatusRuleConditionsField } from '@form-builder/fields/StatusRuleConditionsField';
+import { RedirectRuleConditionsField } from '@form-builder/fields/RedirectRuleConditionsField';
 import { StatusRulesField } from '@form-builder/fields/StatusRulesField';
+import { RedirectRulesField } from '@form-builder/fields/RedirectRulesField';
 import { NestedLayoutField } from '@form-builder/fields/NestedLayoutField';
 import { IntegrationFieldMappingField } from '@form-builder/fields/IntegrationFieldMappingField';
 import { IntegrationRefreshSelectField } from '@form-builder/fields/IntegrationRefreshSelectField';
@@ -153,6 +155,8 @@ Craft.Formie.FormBuilder = function(settings) {
         nextButtonConditions: NextButtonConditionsField,
         statusRuleConditions: StatusRuleConditionsField,
         statusRules: StatusRulesField,
+        redirectRuleConditions: RedirectRuleConditionsField,
+        redirectRules: RedirectRulesField,
         nestedLayout: NestedLayoutField,
         integrationFieldMapping: IntegrationFieldMappingField,
         integrationRefreshSelect: IntegrationRefreshSelectField,